Added the MapCreator Module. Improved the MapViewer. Updated the menus structure.
The Map Creator still not funcional, some of the main features don't
work. Some changes were made on the Map classes in order to implement
it. It uses a MapViewer for displaying the changes on SFML.

The Map Viewer now shows a bar with the continents, the amount of
territories in each one, and it's respective color.